+++ packages/pie-models/src/CommonConfigSettings.ts
@@ -1,0 +1,10 @@
+/**
+ * Config panel settings that are shared across many PIEs
+ */
+export interface CommonConfigSettings {
+  /**
+     * Indicates whether the settings panel wil allow the author to modify settings for partial scoring
+     * @default true
+     */
+    settingsPartialScoring?:  boolean;
+}