+++ packages/extended-text-entry/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.4.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry@3.3.3...@pie-element/extended-text-entry@3.4.0) (2019-02-08)
+
+
+### Features
+
+* json schemas and docs ([6bfc2a9](https://github.com/pie-framework/pie-elements/commit/6bfc2a9))
+* **extended-text-entry:** removed the unnecessary space ([a1dfd90](https://github.com/pie-framework/pie-elements/commit/a1dfd90))
+
+
+
+
+
+## [3.3.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry@3.3.2...@pie-element/extended-text-entry@3.3.3) (2019-01-08)
+
+
+### Bug Fixes
+
+* **general:** bumped the config-ui version for all packages ([159521d](https://github.com/pie-framework/pie-elements/commit/159521d))
+
+
+
+
+
 <a name="3.3.2"></a>
 ## [3.3.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry@3.3.1...@pie-element/extended-text-entry@3.3.2) (2018-10-10)