+++ packages/graph-lines/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines@1.1.4...@pie-element/graph-lines@1.1.5) (2019-01-08)
+
+
+### Bug Fixes
+
+* **general:** bumped the config-ui version for all packages ([159521d](https://github.com/pie-framework/pie-elements/commit/159521d))
+
+
+
+
+
 ## [1.1.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines@1.1.3...@pie-element/graph-lines@1.1.4) (2018-11-15)