+++ packages/inline-choice/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@2.0.5...@pie-element/inline-choice@2.1.0) (2019-02-08)
+
+
+### Features
+
+* json schemas and docs ([6bfc2a9](https://github.com/pie-framework/pie-elements/commit/6bfc2a9))
+
+
+
+
+
+## [2.0.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@2.0.4...@pie-element/inline-choice@2.0.5) (2019-01-08)
+
+
+### Bug Fixes
+
+* **general:** bumped the config-ui version for all packages ([159521d](https://github.com/pie-framework/pie-elements/commit/159521d))
+
+
+
+
+
 <a name="2.0.4"></a>
 ## [2.0.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@2.0.3...@pie-element/inline-choice@2.0.4) (2018-09-20)