+++ packages/multiple-choice/CHANGELOG.md
@@ -3,6 +3,75 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@2.4.0...@pie-element/multiple-choice@2.5.0) (2019-02-08)
+
+
+### Bug Fixes
+
+* **multiple-choice:** fixed a warning ([c8ccb36](https://github.com/pie-framework/pie-elements/commit/c8ccb36))
+* add partial scoring and shuffle prop for multiple-choice ([7577651](https://github.com/pie-framework/pie-elements/commit/7577651))
+
+
+### Features
+
+* **multiple-choice:** added a configurable property for the multiple-choice configure element and fixed some names for the customization properties ([9959eda](https://github.com/pie-framework/pie-elements/commit/9959eda))
+* **multiple-choice:** added a third option for the choices label and also fixed a configuration property name ([2f6d28d](https://github.com/pie-framework/pie-elements/commit/2f6d28d))
+* **multiple-choice:** changed the behaviour when enableShowPrompt is false ([5edaad1](https://github.com/pie-framework/pie-elements/commit/5edaad1))
+* **multiple-choice:** made some change in order to be able to publish a new version ([2be2aae](https://github.com/pie-framework/pie-elements/commit/2be2aae))
+* **multiple-choice:** made some change to the customization properties ([6f0d1b6](https://github.com/pie-framework/pie-elements/commit/6f0d1b6))
+* json schemas and docs ([6bfc2a9](https://github.com/pie-framework/pie-elements/commit/6bfc2a9))
+* **multiple-choice:** updated package version ([8f75d56](https://github.com/pie-framework/pie-elements/commit/8f75d56))
+
+
+
+## 2.4.1 (2019-01-08)
+
+
+### Bug Fixes
+
+* **multiple-choice:** bumped the editable-html version ([220b147](https://github.com/pie-framework/pie-elements/commit/220b147))
+
+
+
+
+
+# [2.4.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@2.3.3...@pie-element/multiple-choice@2.4.0) (2019-01-08)
+
+
+### Bug Fixes
+
+* **tests:** fixed tests for the multiple choice element ([0fa715c](https://github.com/pie-framework/pie-elements/commit/0fa715c))
+
+
+### Features
+
+* **multiple-choice:** moved the theming to the config-ui layout element and removed some unnecessary code ([910e799](https://github.com/pie-framework/pie-elements/commit/910e799))
+* **multiple-choice:** updated snapshot in order to fix tests ([3fb6464](https://github.com/pie-framework/pie-elements/commit/3fb6464))
+* **multiple-choice:** updated the config-ui package and the pie-ui package ([95ae357](https://github.com/pie-framework/pie-elements/commit/95ae357))
+* **multiple-choice:** updated the version of the config-ui ([1d79056](https://github.com/pie-framework/pie-elements/commit/1d79056))
+
+
+
+## 2.3.5 (2018-12-20)
+
+
+### Features
+
+* **multiple-choice:** updated the config-ui package ([bc7aac1](https://github.com/pie-framework/pie-elements/commit/bc7aac1))
+
+
+
+## 2.3.4 (2018-12-17)
+
+
+### Features
+
+* **multiple-choice:** refactored the code of the multiple-choice package in order to use the new layout ([b16b654](https://github.com/pie-framework/pie-elements/commit/b16b654))
+
+
+
+
+
 ## [2.3.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@2.3.2...@pie-element/multiple-choice@2.3.3) (2018-11-14)
 
 **Note:** Version bump only for package @pie-element/multiple-choice