+++ packages/placement-ordering/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@3.2.1...@pie-element/placement-ordering@3.2.3) (2019-02-08)
+
+**Note:** Version bump only for package @pie-element/placement-ordering
+
+
+
+
+
+## [3.2.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@3.2.0...@pie-element/placement-ordering@3.2.1) (2019-01-08)
+
+
+### Bug Fixes
+
+* **general:** bumped the config-ui version for all packages ([159521d](https://github.com/pie-framework/pie-elements/commit/159521d))
+
+
+
+
+
 # [3.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@3.1.0...@pie-element/placement-ordering@3.2.0) (2018-11-12)