+++ packages/point-intercept/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/point-intercept@3.1.4...@pie-element/point-intercept@3.1.6) (2019-02-08)
+
+**Note:** Version bump only for package @pie-element/point-intercept
+
+
+
+
+
+## [3.1.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/point-intercept@3.1.3...@pie-element/point-intercept@3.1.4) (2019-01-08)
+
+
+### Bug Fixes
+
+* **general:** bumped the config-ui version for all packages ([159521d](https://github.com/pie-framework/pie-elements/commit/159521d))
+
+
+
+
+
 <a name="3.1.3"></a>
 ## [3.1.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/point-intercept@3.1.2...@pie-element/point-intercept@3.1.3) (2018-09-20)