+++ packages/select-text/configure/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-configure@2.2.2...@pie-element/select-text-configure@2.2.3) (2019-01-08)
+
+
+### Bug Fixes
+
+* **general:** bumped the config-ui version for all packages ([159521d](https://github.com/pie-framework/pie-elements/commit/159521d))
+
+
+
+
+
 ## [2.2.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-configure@2.2.1...@pie-element/select-text-configure@2.2.2) (2018-11-12)